[CHORE]: ErrorHandling Protocol로 변경
